Fix random sampling #9

+++ dogm/src/kernel/resampling.cu
@@ -31,7 +31,7 @@ SOFTWARE.
 
 #include <thrust/binary_search.h>
 
-void calc_resampled_indices(thrust::device_vector<float>& joint_weight_accum, thrust::device_vector<int>& rand_array,
+void calc_resampled_indices(thrust::device_vector<float>& joint_weight_accum, thrust::device_vector<float>& rand_array,
 	thrust::device_vector<int>& indices)
 {
 	thrust::device_vector<float> norm_weight_accum(joint_weight_accum.size());
@@ -43,13 +43,13 @@ void calc_resampled_indices(thrust::device_vector<float>& joint_weight_accum, th
 	});
 
 	float norm_max = norm_weight_accum.back();
-	int rand_max = rand_array.back();
+	float rand_max = rand_array.back();
 
-	printf("Norm: %f, Rand: %d\n", norm_max, rand_max);
+	printf("Norm: %f, Rand: %f\n", norm_max, rand_max);
 
 	if (norm_max != rand_max)
 	{
-		norm_weight_accum.back() = static_cast<float>(rand_max);
+		norm_weight_accum.back() = rand_max;
 	}
 
 	// multinomial sampling
